### 2019-01-19 #1187 @whedon
[PRE REVIEW]: areal: An R package for areal weighted interpolation
**Submitting author:** @chris-prener (<a href="http://orcid.org/0000-0002-4310-9888">Christopher Prener</a>)
**Repository:** <a href="https://github.com/slu-openGIS/areal" target ="_blank">https://github.com/slu-openGIS/areal</a>
**Version:** v0.1.3
