### 2019-01-19 #1187 @chris-prener
> Looks like we failed to compile the PDF

I'm wondering if this failed because of [the use of backticks in the heading](https://github.com/slu-openGIS/areal/blob/master/paper/paper.md#the-areal-r-package)? If so, I can remove and re-push...
