### 2019-01-20 #1187 @arfon
This failure is due to a bug we recently introduced into our LaTeX template. I'll fix this shortly and let you know when it's updated.
