### 2019-01-21 #1187 @chris-prener
>👋 Hi @chris-prener, thanks for your submission! Do you have any recommendations for reviewers?

Hi @lheagy - the following three people from google doc all list GIS and R as areas/languages of interest: nuest, jsta, sjsrey.
