### 2019-01-27 #1187 @lheagy
👋 Hi @nuest, @sjsrey, would you be willing to review this submission for JOSS?
