### 2019-01-28 #1187 @sjsrey
Hi Lindsey,

I would be happy to help - with the one caveat that I can't start on this
for another two weeks as I have several other reviews in my queue that I am
committed to.

s.
Sergio (Serge) Rey
Founding Director, Center for Geospatial Sciences
Professor, School of Public Policy
University of California, Riverside
sergio.rey@ucr.edu

On Sun, Jan 27, 2019 at 10:41 AM Lindsey Heagy <notifications@github.com>
wrote:

> 👋 Hi @nuest <https://github.com/nuest>, @sjsrey
> <https://github.com/sjsrey>, would you be willing to review this
> submission for JOSS?
>
