### 2019-01-29 #1187 @lheagy
Thanks for your reply @nuest.

:wave: Hi @edzer, would you be willing to review this submission for JOSS?
