### 2019-01-30 #1187 @edzer
Sure. Could you please point me to the guidelines for reviewers?
