### 2019-01-30 #1187 @lheagy
Excellent, thank you @edzer! The guidelines for the review process are here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html. Shortly, I will close this issues and open another that contains a checklist that will help guide the review process. Please let me know if you have any questions or if I can provide further clarification.
