### 2019-01-30 #1221 @lheagy
:wave: Many thanks @sjsrey, @edzer for being willing to review! In the main thread above, there is a checklist for you both to help guide the review. If you haven't previously reviewed for JOSS, please be sure to accept the invite: https://github.com/openjournals/joss-reviews/invitations. This will allow you to check off the items in the checklist. If possible, we would appreciate receiving your review within the next 2 weeks.

Please let me know if you have any questions or if I can provide clarification on anything.
