### 2019-02-04 #1221 @sjsrey
@chris-prener when installing via
```
> install.packages("areal")
```
 I'm getting:
```
ERROR: dependencies ‘lwgeom’, ‘sf’ are not available for package ‘areal’
  * removing ‘/home/serge/anaconda3/lib/R/library/areal’

The downloaded source packages are in
        ‘/tmp/RtmpBEpBpJ/downloaded_packages’
Updating HTML index of packages in '.Library'
Making 'packages.html' ... done
Warning messages:
1: In install.packages("areal") :
  installation of package ‘units’ had non-zero exit status
2: In install.packages("areal") :
  installation of package ‘sf’ had non-zero exit status
3: In install.packages("areal") :
  installation of package ‘lwgeom’ had non-zero exit status
4: In install.packages("areal") :
  installation of package ‘areal’ had non-zero exit status
> library(areal)
Error in library(areal) : there is no package called ‘areal’
>
```

I realize this might be due to using R installed via conda. But on Ubuntu 18.04 I'm getting the same error(s) on a clean install of R from the official repositories:
```
* DONE (dplyr)
ERROR: dependencies ‘lwgeom’, ‘sf’ are not available for package ‘areal’
* removing ‘/home/serge/R/x86_64-pc-linux-gnu-library/3.4/areal’

The downloaded source packages are in
        ‘/tmp/Rtmpb2kPO5/downloaded_packages’
Warning messages:
1: In install.packages("areal") :
  installation of package ‘units’ had non-zero exit status
2: In install.packages("areal") :
  installation of package ‘sf’ had non-zero exit status
3: In install.packages("areal") :
  installation of package ‘lwgeom’ had non-zero exit status
4: In install.packages("areal") :
  installation of package ‘areal’ had non-zero exit status
> library(areal)
Error in library(areal) : there is no package called ‘areal’
```
