### 2019-02-13 slu-openGIS/areal#18 @edzer
Namespaces in Imports field not imported from
When I run R CMD check on the package, I see
```r
* checking dependencies in R code ... NOTE
Namespaces in Imports field not imported from:
  ‘lwgeom’ ‘tibble’
```
suggesting that these two packages can be removed from `Imports:`
https://github.com/openjournals/joss-reviews/issues/1221
