### 2019-02-13 #1221 @edzer
This is a nice package that provides relatively limited functionality over `sf::st_interpolate_aw` (which I wrote), but that explains and illustrates the whole areal interpolation process very well. Since JOSS does not demand new functionality, it is a nice addition to the existing package ecosystem around `sf`. I recommend that the authors address the following issues:

* The pdf of the paper shows R code on page 2, but this code is hidden where it runs off the page at the side.
* The manuscript claims it provides functions that "validate data suitability for areal interpolation", but does not explain what this validation does. What does it do?
* given that the package does little more than `sf::st_interpolate_aw`, and given that that function was available first, the added value of the packages would become more clear when it would
     * mention the name of that _function_ (rather than refer to "the existing functionality in `sf`"),
     * show in a side-by-side comparison that `areal::aw_interpolate` and `sf::st_interpolate_aw` give identical results for an intensive an an extensive variable
     * give an example where `areal::aw_interpolate` does something different when there is a difference in areas, and show how large the effect of doing this is
* the correct reference for _Pebesma, E. (2018)_ is found [here](https://journal.r-project.org/archive/2018/RJ-2018-009/index.html)
* change the wording of "black box", as this is not a black box. Closed source software is a black box
* the `areal-weighted-interpolation` vignette mentions "Spatially intensive operations are used when the data to be interpolated are a ratio." I'm afraid it is not this simple: a counter example is is CO2 emissions measured in tons/year: an extensive variable when interpolated spatially.
