### 2019-02-15 slu-openGIS/areal#18 @chris-prener
Thanks @edzer - we want to make sure `lwgeom` is installed because we use `sf::st_area` in the package. So while we don't import anything directly, it is a necessary dependency that isn't installed automatically by `sf`. At least that is how I am thinking about it - am I off base there?

With `tibble` - my understanding is that if a package prints an object as a tibble but the user does not have `tibble` installed, there are issues with printing the object. So therefore safter to require `tibble` as a dependency? Again, would love your perspective here.

Thanks for the feedback!
