### 2019-02-15 slu-openGIS/areal#18 @edzer
`st_area` is in `sf`, and `sf` only class lwgeom stuff if the data are in geodetic coordinates, which is often not the case, so you don't actually need it, and `sf` will tell you when you need it. `tibble` is a dependency of `dplyr`, so comes automatically when you import `dplyr`.

If CRAN has no problems with what you do, then I also don't.
