### 2019-02-15 slu-openGIS/areal#18 @chris-prener
We put it in when we were testing the package and were prompted to install `lwgeom` by `sf`. I'll revisit whether it is necessary, though. CRAN had no problems.
