### 2019-02-16 #1221 @sjsrey
Well designed and documented package that extends important functionality for areal interpolation. A couple of points the authors may want to consider - these are not show stoppers by any means just food for thought and perhaps future extensions:

* Scaling: the current examples are well crafted and illustrate the core functionality. At the same time `n` is fairly modest here and one wonders if these methods were going to be used for larger `n` problems, or applied iteratively over [many cities ](https://spatial.ucr.edu/project_neighborhood.html), would there be any bottlenecks users should be aware of?
* The distinction between extensive and intensive variables is handled nicely. One case that might also be considered is the choice between interpolation of an intensive variable directly versus deriving the intensive estimates as a ratio of two extensive variables that have been estimated: `pct_black = black/total`, for example could be based on first estimating `black`, and `total`, or it could be treated as an intensive variable that is estimated in one step.. Related to this is that there are inherent constraints that are likely violated when treating the intensive and extensive variables independently in certain cases. For example, in a simple world where where 'total = black + white`, this implies `pct_black + pct_white =1`. If the intensive and extensive variables involved are estimated independently, this constraint is likely to be violated.
*  The `weight` argument names are a bit confusing. `sum`, to me, would seem to a better name for the case where intersection  between the target areas and a source area do not exhaust the source area. In other words, use the sum of the intersection area over the source area to form the weight. `total`, to me, seems to imply one would want to allocate the total value of the source area over the target areas.
