### 2019-02-17 #1221 @chris-prener
Thanks @sjsrey - I appreciate your feedback. We can absolutely address your last point to clarify `sum` and `total`. We can also clarify point two a bit in the documentation.

Re: your first bullet - I imagine there would be speed issues with a large `n`... a benchmark where we interpolate from state to counties might possibly illustrate speed constraints?

In any event - we can address these points in an updated draft. Will post here when the paper manuscript and associated documentation for the package has been updated. @lheagy - aside from addressing reviewer concerns, are there any other steps we need to take at this stage?
