### 2019-03-22 70dd36b
And actually add the st_geometry_type call.
