### 2019-03-22 9adc072
test for GEOMETRYCOLLECTION cast to MULTIPOLYGON
