### 2019-03-22 96971fa
remove st_cast call, update unit tests
