+++ R/aw_intersect.R
@@ -65,9 +65,9 @@ aw_intersect <- function(.data, source, areaVar) {
   intersection %>%
     aw_area(areaVar = !!areaVarQ) -> intersection
 
+  # if a geometry collection is returned, extract it
   if(any(grepl("GEOMETRY", sf::st_geometry_type(intersection)))) {
-    intersection <- sf::st_collection_extract(intersection) %>%
-      sf::st_cast("MULTIPOLYGON")
+    intersection <- sf::st_collection_extract(intersection)
   }
 
   # return output
