### 2019-03-22 slu-openGIS/areal#18 @chris-prener
have removed `lwgeom` (shouldn't be needed since projected coords required by functions) and `tibble` (which should be installed along with `dplyr`)
