### 2019-03-22 6af5fac
enable appveyor
