+++ appveyor.yml
@@ -18,7 +18,7 @@ cache:
 environment:
   global:
     WARNINGS_ARE_ERRORS: 1
-#    USE_RTOOLS: true
+    USE_RTOOLS: true
 
   matrix:
     - R_VERSION: devel
