+++ paper/paper.bib
@@ -47,10 +47,16 @@
 }
 
 @article{pebesma2018simple,
-  title={Simple Features for R: Standardized Support for Spatial Vector Data},
-  author={Pebesma, Edzer},
-  journal={The R Journal},
-  year={2018}
+  author = {Edzer Pebesma},
+  title = {{Simple Features for R: Standardized Support for Spatial
+          Vector Data}},
+  year = {2018},
+  journal = {{The R Journal}},
+  doi = {10.32614/RJ-2018-009},
+  url = {https://doi.org/10.32614/RJ-2018-009},
+  pages = {439--446},
+  volume = {10},
+  number = {1}
 }
 
 @online{prener2019areal,
