### 2019-03-23 fb05119
address issues with two vectors in mixed interpolations
