### 2019-03-23 0cae023
update paper manuscript, vignettes
