### 2019-03-23 bdbbfc4
add notes about additional bug fixes
