### 2019-03-23 f8a13d7
update NEWS with new version
