+++ vignettes/areal-weighted-interpolation.Rmd
@@ -298,6 +298,24 @@ race %>%
     TOTAL_PCT = WHITE_PCT+BLACK_PCT
   ) -> constrainedData
 
+# interpolate
+result <- aw_interpolate(ar_stl_wards, tid = WARD, 
+               source = constrainedData, sid = GEOID, 
+               weight = "sum", output = "tibble", 
+               intensive = c("WHITE_PCT", "BLACK_PCT", "TOTAL_PCT"),
+               extensive = c("TOTAL", "WHITE_E", "BLACK_E"))
+
+# calculate new percentages
+result %>%
+  mutate(
+    WHITE_PCT_2 = WHITE_E/TOTAL,
+    BLACK_PCT_2 = BLACK_E/TOTAL,
+    TOTAL_PCT_2 = WHITE_PCT_2+BLACK_PCT_2
+  ) -> result
+
+# display
+result %>%
+  select(WHITE_PCT, WHITE_PCT_2, BLACK_PCT, BLACK_PCT_2, TOTAL_PCT, TOTAL_PCT_2)
 ```
 
 Note that there are a number of points of departure between the data interpolated as intensive values (`WHITE_PCT`, `BLACK_PCT`) and those that were interpolated as count data (i.e. extensive values) and then converted to intensive variables (`WHITE_PCT_2` and `BLACK_PCT_2`). 
