+++ paper/appendix.md
@@ -1,0 +1,316 @@
+Appendix for JOSS Paper
+================
+Christopher Prener, Ph.D.
+(March 23, 2019)
+
+## Introduction
+
+This document contains replication code for the examples provided in our
+*Journal of Open Source Software* manuscript.
+
+## Dependencies
+
+This notebook requires
+
+``` r
+# primary package
+library(areal)
+
+# tidyverse packages
+library(dplyr)
+```
+
+    ## 
+    ## Attaching package: 'dplyr'
+
+    ## The following objects are masked from 'package:stats':
+    ## 
+    ##     filter, lag
+
+    ## The following objects are masked from 'package:base':
+    ## 
+    ##     intersect, setdiff, setequal, union
+
+``` r
+# spatial packages
+library(sf)
+```
+
+    ## Linking to GEOS 3.6.1, GDAL 2.1.3, PROJ 4.9.3
+
+``` r
+library(tidycensus)
+library(tigris)
+```
+
+    ## To enable 
+    ## caching of data, set `options(tigris_use_cache = TRUE)` in your R script or .Rprofile.
+
+    ## 
+    ## Attaching package: 'tigris'
+
+    ## The following object is masked from 'package:graphics':
+    ## 
+    ##     plot
+
+``` r
+# other packages
+library(gridExtra)
+```
+
+    ## 
+    ## Attaching package: 'gridExtra'
+
+    ## The following object is masked from 'package:dplyr':
+    ## 
+    ##     combine
+
+``` r
+library(microbenchmark)
+library(testthat)
+```
+
+    ## 
+    ## Attaching package: 'testthat'
+
+    ## The following object is masked from 'package:dplyr':
+    ## 
+    ##     matches
+
+## Comparisons with `sf`
+
+### Produce Estimates
+
+First, we’ll create three spatially extensive estimates for comparison.
+Two will use the `areal` package, varying the type of weight applied to
+the estimate:
+
+``` r
+# areal package, spatially extensive using total
+areal_exT <- aw_interpolate(ar_stl_wards, tid = WARD, source = ar_stl_race, sid = GEOID,
+               weight = "total", output = "tibble", extensive = "TOTAL_E")
+
+# areal package, spatially extensive using sum
+areal_exS <- aw_interpolate(ar_stl_wards, tid = WARD, source = ar_stl_race, sid = GEOID,
+                            weight = "sum", output = "tibble", extensive = "TOTAL_E")
+```
+
+Next, we’ll replicate the process using `sf`:
+
+``` r
+# sf package, spatially extensive
+sf_ex <- st_interpolate_aw(ar_stl_race["TOTAL_E"], ar_stl_wards, extensive = TRUE)
+```
+
+    ## Warning in st_interpolate_aw.sf(ar_stl_race["TOTAL_E"], ar_stl_wards,
+    ## extensive = TRUE): st_interpolate_aw assumes attributes are constant over
+    ## areas of x
+
+We’ll also produce a spatially intensive estimate using `areal`:
+
+``` r
+# areal package, spatially intensive
+areal_in <- aw_interpolate(ar_stl_wards, tid = WARD, source = ar_stl_asthma, sid = GEOID,
+                            weight = "sum", output = "tibble", intensive = "ASTHMA")
+```
+
+And finally, we’ll replicate the spatially intensive estimate using
+`sf`:
+
+``` r
+# sf package, spatially intensive
+sf_in <- st_interpolate_aw(ar_stl_asthma["ASTHMA"], ar_stl_wards, extensive = FALSE)
+```
+
+    ## Warning in st_interpolate_aw.sf(ar_stl_asthma["ASTHMA"], ar_stl_wards,
[…]

[…]
