### 2019-03-29 #1221 @lheagy
Many thanks for your thorough response @chris-prener!

@edzer, @sjsrey would you be willing to take a look through @chris-prener's response? Please check off items in the checklist that have been adequately addressed, and if there are any remaining items that you think should be changed before acceptance, please leave comments here. Thank you both!
