### 2019-04-03 slu-openGIS/areal#20 @lheagy
There are also a few missing doi's in the references

- https://doi.org/10.1016/j.compenvurbsys.2005.07.005 may be missing for title: Rapid facilitation of dasymetric-based population interpolation by means of raster pixel maps
- https://doi.org/10.1080/00182494.1973.10112670 may be missing for title: The linkage of data describing overlapping geographical units
- https://doi.org/10.2747/1548-1603.49.5.644 may be missing for title: The development of an areal interpolation ArcGIS extension and a comparative study
