### 2019-04-03 #1221 @lheagy
@chris-prener: I left a few comments for grammar in the paper in slu-openGIS/areal#20, and it looks like there are a few missing doi's in the references. Would you please take a look at these and ping here when you are done? Thanks!
