### 2019-04-04 slu-openGIS/areal#20 @chris-prener
Thanks @lheagy - will get back with changes this weekend.
