### 2019-04-16 #1221 @lheagy
:wave: Hi @chris-prener just checking in - have you had a chance to take a look at the suggestions in slu-openGIS/areal#20?
