### 2019-04-20 73c043a
add missing doi numbers
