+++ paper/paper.bib
@@ -17,6 +17,7 @@
   title={Spatial interpolation methods: a review},
   author={Lam, Nina Siu-Ngan},
   journal={The American Cartographer},
+  doi = {10.1559/152304083783914958},
   volume={10},
   number={2},
   pages={129--150},
@@ -28,6 +29,7 @@
   title={Rapid facilitation of dasymetric-based population interpolation by means of raster pixel maps},
   author={Langford, Mitchel},
   journal={Computers, Environment and Urban Systems},
+  doi = {10.1016/j.compenvurbsys.2005.07.005},
   volume={31},
   number={1},
   pages={19--32},
@@ -39,6 +41,7 @@
   title={The linkage of data describing overlapping geographical units},
   author={Markoff, John and Shapiro, Gilbert},
   journal={Historical Methods Newsletter},
+  doi = {10.1080/00182494.1973.10112670},
   volume={7},
   number={1},
   pages={34--46},
@@ -64,13 +67,15 @@
   title = {areal: Areal Weighted Interpolation},
   year = 2019,
   url = {https://CRAN.R-project.org/package=areal },
-  urldate = {2018-12-20}
+  urldate = {2018-12-20},
+  doi = {10.5281/zenodo.2603540}
 }
 
 @article{qiu2012development,
   title={The development of an areal interpolation ArcGIS extension and a comparative study},
   author={Qiu, Fang and Zhang, Caiyun and Zhou, Yuhong},
   journal={GIScience \& Remote Sensing},
+  doi = {10.2747/1548-1603.49.5.644},
   volume={49},
   number={5},
   pages={644--663},
