### 2019-04-20 slu-openGIS/areal#20 @chris-prener
All changes have been made and committed to the master branch @lheagy
