### 2019-04-20 #1221 @chris-prener
hey @lheagy - thanks for your patience! just made all of those suggested changes and added the DOI numbers. I also found an additional DOI that whedon did not, and added the package's DOI as well.
