### 2019-04-23 #1221 @lheagy
many thanks @chris-prener. Please have another look at the compiled pdf above and double-check that all looks good (nothing jumped out at me). If it is fine, then please create a new release and archive the software on [zenodo](https://zenodo.org/). The title and author list on the zenodo archive should match the paper. Once you have done this, please post the version number and doi here and we can proceed with publishing 🎉
