### 2019-04-26 308e22b
minor edits to paper manuscript
