+++ paper/paper.bib
@@ -49,6 +49,14 @@
   publisher={Taylor \& Francis}
 }
 
+@Manual{olaf2018micro,
+    title = {microbenchmark: Accurate Timing Functions},
+    author = {Olaf Mersmann},
+    year = {2018},
+    note = {R package version 1.4-6},
+    url = {https://CRAN.R-project.org/package=microbenchmark},
+  }
+
 @article{pebesma2018simple,
   author = {Edzer Pebesma},
   title = {{Simple Features for R: Standardized Support for Spatial
