### 2019-04-26 da4e452
update version number for release, add citation file
