### 2019-04-26 #1221 @chris-prener
hi @lheagy - I corrected a few additional typos and added a citation for the `microbenchmark` package. The most recent paper has been pushed to the repo and released. The zendo repository has been updated with a `.zenodo.json` file to make sure it is formatted the way you described above.
