### 2019-04-26 #1221 @chris-prener
@lheagy - everything looks good. Is there a way to force the sample `R` code onto the next page so it isn't broken up by the page break?
