### 2019-05-01 #1221 @lheagy
Thanks @chris-prener.

@arfon: is there an easy way to keep a code-block together in the paper? (see code split over page 2-3 here: https://github.com/openjournals/joss-papers/blob/joss.01221/joss.01221/10.21105.joss.01221.pdf)
