### 2019-05-01 #1221 @arfon
> @arfon: is there an easy way to keep a code-block together in the paper? (see code split over page 2-3 here: https://github.com/openjournals/joss-papers/blob/joss.01221/joss.01221/10.21105.joss.01221.pdf)

Perhaps a `\newpage` would help here? I _think_ this works with Pandoc/Markdown.
