### 2019-05-01 5a4e6f8
update NEWS, DESCRIPTION, and pkgdown site
