### 2019-05-01 51f5139
delete ar_validate subfunction docs
