+++ R/ar_validate.R
@@ -26,6 +26,7 @@
 #' ar_validate(source = ar_stl_asthma, target = ar_stl_wards, varList = "ASTHMA", verbose = TRUE)
 #'
 #' @importFrom glue glue
+#' @importFrom sf st_crs st_is_longlat
 #'
 #' @export
 ar_validate <- function(source, target, varList, method = "aw", verbose = FALSE){
@@ -117,21 +118,19 @@ ar_validate <- function(source, target, varList, method = "aw", verbose = FALSE)
 
 }
 
-#' Lite Version of Validation for aw_preview_weights
-#'
-#' @description \code{aw_validate_preview} is designed to be called by
-#'     \code{aw_preview_weights} before the weights are calculated. It
-#'     lacks the variable validation functionality of \code{ar_validate}.
-#'
-#' @param source A \code{sf} object with data to be interpolated
-#' @param target A \code{sf} object that data should be interpolated to
-#'
-#' @return If \code{verbose} is \code{FALSE}, a logical scalar is returned that is \code{TRUE}
-#'     is all tests are passed and \code{FALSE} if one or more tests is failed. If \code{verbose}
-#'     is \code{TRUE}, a tibble with detailed test results is returned.
-#'
-#' @importFrom glue glue
-#'
+# Lite Version of Validation for aw_preview_weights
+#
+# @description \code{aw_validate_preview} is designed to be called by
+#     \code{aw_preview_weights} before the weights are calculated. It
+#     lacks the variable validation functionality of \code{ar_validate}.
+#
+# @param source A \code{sf} object with data to be interpolated
+# @param target A \code{sf} object that data should be interpolated to
+#
+# @return If \code{verbose} is \code{FALSE}, a logical scalar is returned that is \code{TRUE}
+#     is all tests are passed and \code{FALSE} if one or more tests is failed. If \code{verbose}
+#     is \code{TRUE}, a tibble with detailed test results is returned.
+#
 aw_validate_preview <- function(source, target){
 
   # store results from primary validate subfunctions
@@ -172,16 +171,16 @@ aw_validate_preview <- function(source, target){
 
 }
 
-#' Testing for sf object status for source and target data
-#'
-#' @description \code{ar_validate_sf} conducts a logic test for shared coordinate
-#'     coordinate systems, which are a requirement for interpolation.
-#'
-#' @param source A \code{sf} object with data to be interpolated
-#' @param target A \code{sf} object that data should be interpolated to
-#'
-#' @return A logical scalar; if \code{TRUE}, the test is passed.
-#'
+# Testing for sf object status for source and target data
+#
+# @description \code{ar_validate_sf} conducts a logic test for shared coordinate
+#     coordinate systems, which are a requirement for interpolation.
+#
+# @param source A \code{sf} object with data to be interpolated
+# @param target A \code{sf} object that data should be interpolated to
+#
+# @return A logical scalar; if \code{TRUE}, the test is passed.
+#
 ar_validate_sf <- function(source, target){
 
   # identify sf object in class
@@ -205,18 +204,16 @@ ar_validate_sf <- function(source, target){
 
 }
 
-#' Testing for shared coordinates for source and target data
-#'
-#' @description \code{awrvalidate_crs} conducts a logic test for shared coordinate
-#'     coordinate systems, which are a requirement for interpolation.
-#'
-#' @param source A \code{sf} object with data to be interpolated
-#' @param target A \code{sf} object that data should be interpolated to
-#'
-#' @return A logical scalar; if \code{TRUE}, the test is passed.
-#'
-#' @importFrom sf st_crs
-#'
+# Testing for shared coordinates for source and target data
+#
[…]
