+++ man/ar_validate_longlat.Rd
@@ -1,18 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/ar_validate.R
-\name{ar_validate_longlat}
-\alias{ar_validate_longlat}
-\title{Testing for type of coordinates}
-\usage{
-ar_validate_longlat(.data)
-}
-\arguments{
-\item{.data}{A sf object}
-}
-\value{
-A logical scalar; if \code{TRUE}, the test is passed
-}
-\description{
-\code{ar_validate_longlat} conducts a logic test for
-    whether or not the data are in planar format.
-}
