+++ man/ar_validate_vars_conflict.Rd
@@ -1,21 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/ar_validate.R
-\name{ar_validate_vars_conflict}
-\alias{ar_validate_vars_conflict}
-\title{Testing for Variable Conflicts in Target}
-\usage{
-ar_validate_vars_conflict(.data, varList)
-}
-\arguments{
-\item{.data}{A sf object}
-
-\item{varList}{A vector of variables to be created}
-}
-\value{
-A logical scalar; if \code{TRUE}, the test is passed
-}
-\description{
-\code{ar_validate_vars_conflict} conducts a logic test for
-    whether or not any of the variables to be created in the target
-    data already exist as named columns.
-}
