+++ man/ar_validate_vars_exist.Rd
@@ -1,20 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/ar_validate.R
-\name{ar_validate_vars_exist}
-\alias{ar_validate_vars_exist}
-\title{Testing for Variables Existing in Source}
-\usage{
-ar_validate_vars_exist(.data, varList)
-}
-\arguments{
-\item{.data}{A sf object}
-
-\item{varList}{A vector of variables assumed to exist.}
-}
-\value{
-A logical scalar; if \code{TRUE}, the test is passed
-}
-\description{
-\code{ar_validate_vars_exist} conducts a logic test for
-    whether or not all variables exist in the source data.
-}
