+++ man/aw_validate_preview.Rd
@@ -1,23 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/ar_validate.R
-\name{aw_validate_preview}
-\alias{aw_validate_preview}
-\title{Lite Version of Validation for aw_preview_weights}
-\usage{
-aw_validate_preview(source, target)
-}
-\arguments{
-\item{source}{A \code{sf} object with data to be interpolated}
-
-\item{target}{A \code{sf} object that data should be interpolated to}
-}
-\value{
-If \code{verbose} is \code{FALSE}, a logical scalar is returned that is \code{TRUE}
-    is all tests are passed and \code{FALSE} if one or more tests is failed. If \code{verbose}
-    is \code{TRUE}, a tibble with detailed test results is returned.
-}
-\description{
-\code{aw_validate_preview} is designed to be called by
-    \code{aw_preview_weights} before the weights are calculated. It
-    lacks the variable validation functionality of \code{ar_validate}.
-}
