### 2019-05-01 278f441
remove aw_interpolate subfunction docs
