+++ man/aw_strip_df.Rd
@@ -1,24 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/aw_interpolate.R
-\name{aw_strip_df}
-\alias{aw_strip_df}
-\title{Strip dataframe of all non-essential variables}
-\usage{
-aw_strip_df(.data, id, value)
-}
-\arguments{
-\item{.data}{A \code{sf} object}
-
-\item{id}{A given source id field}
-
-\item{value}{Optional; the variable that estimations will be based on}
-}
-\value{
-A \code{sf} object with only the \code{id} and, if provided, the
-    \code{value} column as well.
-}
-\description{
-\code{aw_strip_df} is called by \code{aw_interpolate}. It
-    strips \code{sf} objects of nonessential variables but keeps
-    variables listed in parameters.
-}

[…]
