### 2019-05-01 a3a94e8
modify docs for aw_intersect
