+++ man/aw_area.Rd
@@ -1,20 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/aw_intersect.R
-\name{aw_area}
-\alias{aw_area}
-\title{Calculate area}
-\usage{
-aw_area(.data, areaVar)
-}
-\arguments{
-\item{.data}{A \code{sf} object that data should be interpolated to}
-
-\item{areaVar}{The name of the new area variable to be calculated.}
-}
-\value{
-A \code{sf} object with the new area field.
-}
-\description{
-Calculate the area of a feature in the units of the current
-    coordinate system. This is called by \code{aw_intersect}.
-}
