+++ .travis.yml
@@ -2,7 +2,7 @@
 
 language: R
 sudo: required
-cache: packages
+# cache: packages
 
 r:
   - oldrel
