### 2019-05-02 c21e1f4
remove preview weight subfunction docs
