+++ man/aw_calculate_weight.Rd
@@ -1,21 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/aw_preview_weights.R
-\name{aw_calculate_weight}
-\alias{aw_calculate_weight}
-\title{Caclulate Weights}
-\usage{
-aw_calculate_weight(.data, source, id, item)
-}
-\arguments{
-\item{.data}{A \code{sf} object that data should be interpolated to (this is referred
-to as the \code{target} elsewhere in the package).}
-
-\item{source}{A \code{sf} object with data to be interpolated}
-
-\item{id}{A unique identification number in either the source or target data}
-
-\item{item}{One of \code{"extensive_sum"}, \code{"extensive_total"}, or \code{"intensive"}}
-}
-\description{
-Subfunction of aw_preview_weight for calculating individual weights
-}
