+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: areal
 Type: Package
 Title: Areal Weighted Interpolation
-Version: 0.1.4.2
+Version: 0.1.4.3
 Authors@R: c(
     person("Christopher", "Prener", , "chris.prener@slu.edu", c("aut", "cre"),
     comment = c(ORCID = "0000-0002-4310-9888")),
