### 2019-05-04 #1221 @chris-prener
the zenodo doi for the latest release is `10.5281/zenodo.2667289`; the master doi that always resolves to the latest version is `10.5281/zenodo.2433868`.
