### 2019-05-04 #1221 @kyleniemeyer
@chris-prener we do want to use the specific DOI for the accepted release, since you could theoretically submit a later JOSS paper for a future version
