### 2019-05-05 #1221 @lheagy
:wave: @openjournals/joss-eics: this submission is ready to be accepted 🎉
