### 2019-05-16 #1221 @kyleniemeyer
@chris-prener some minor edits for the paper:
- could you add full affiliation details (i.e., location) for the authors?
- the first paragraph ends in a comma
