### 2019-05-16 410046d
tweaks to JOSS manuscript
