### 2019-05-16 #1221 @chris-prener
@kyleniemeyer and @lheagy - made the two changes Kyle requested, and tagged a new released with the updated paper manuscript. The Zenodo DOI for the newest release is - 10.5281/zenodo.2857598
