### 2019-05-16 #1221 @kyleniemeyer
@chris-prener sorry, looks like Zenodo is having some database issues right now, so I'm going to hold off moving forward until we can confirm the archive. Should be resolved within the day, I hope.
