### 2019-05-19 #1221 @chris-prener
@kyleniemeyer and @lheagy - looks like everything is back up on the Zenodo end. the DOI is resolving correctly and Zenodo has the correct version of the updated software.
