### 2019-06-03 #1221 @chris-prener
@lheagy or @arfon - one follow-up question - does Google Scholar index JOSS publications right now? Googled around but didn't seem to find a clear answer...
