### 2019-06-03 #1221 @arfon
> @lheagy or @arfon - one follow-up question - does Google Scholar index JOSS publications right now? Googled around but didn't seem to find a clear answer...

Yes, but based on historical performance, it can take a month or two for them to show up.
